Add files via upload
INI file in JSON format.  Anything that isn't hardcoded can be changed here.  I appreciate all of the people who will say nothing should be hardcoded, but this is a living repo, some shit is going to be hardcoded even though I know better because I don't care and it's going to be fixed in the next version.  I'll try to comment all magic numbers and hardcoded values where I can, but if you find one I missed, you are right.